Refactored helper class for picker plus WIP test
