+++ viewer/packages/pointclouds/src/potree-three-loader/tree/PointCloudOctreePicker.ts
@@ -1,61 +1,19 @@
 import {
-  BufferAttribute,
   Camera,
-  Color,
-  LinearFilter,
-  NearestFilter,
-  NoBlending,
-  Points,
   Ray,
-  RGBAFormat,
-  Scene,
-  Sphere,
   Vector3,
-  Vector4,
   WebGLRenderer,
-  WebGLRenderTarget
 } from 'three';
-import { COLOR_BLACK, DEFAULT_PICK_WINDOW_SIZE } from '../rendering/constants';
-import { ClipMode, PointCloudMaterial, PotreePointColorType } from '../rendering';
+import { DEFAULT_PICK_WINDOW_SIZE } from '../rendering/constants';
 import { PointCloudOctree } from './PointCloudOctree';
-import { IPointCloudTreeNode } from './IPointCloudTreeNode';
-import { PickPoint, PointCloudHit } from '../types/types';
+import { PickPoint } from '../types/types';
 import { clamp } from '../utils/math';
+import { IPickState, PickParams, PointCloudOctreePickerHelper } from './PointCloudOctreePickerHelper';
 
-export interface PickParams {
-  pickWindowSize: number;
-  pickOutsideClipRegion: boolean;
-  /**
-   * If provided, the picking will use this pixel position instead of the `Ray` passed to the `pick`
-   * method.
-   */
-  pixelPosition: Vector3;
-  /**
-   * Function which gets called after a picking material has been created and setup and before the
-   * point cloud is rendered into the picking render target. This gives applications a chance to
-   * customize the renderTarget and the material.
-   *
-   * @param material The pick material.
-   * @param renterTarget The render target used for picking.
-   */
-  onBeforePickRender: (material: PointCloudMaterial, renterTarget: WebGLRenderTarget) => void;
-}
-
-interface IPickState {
-  renderTarget: WebGLRenderTarget;
-  material: PointCloudMaterial;
-  scene: Scene;
-}
-
-interface RenderedNode {
-  node: IPointCloudTreeNode;
-  octree: PointCloudOctree;
-}
+export { PickParams };
 
 export class PointCloudOctreePicker {
   private static readonly helperVec3 = new Vector3();
-  private static readonly helperSphere = new Sphere();
-  private static readonly clearColor = new Color();
   private pickState: IPickState | undefined;
 
   dispose(): void {
@@ -76,14 +34,14 @@ export class PointCloudOctreePicker {
       return null;
     }
 
-    const pickState = this.pickState ? this.pickState : (this.pickState = PointCloudOctreePicker.getPickState());
+    const pickState = this.pickState ? this.pickState : (this.pickState = PointCloudOctreePickerHelper.getPickState());
 
     const pickMaterial = pickState.material;
 
     const pixelRatio = renderer.getPixelRatio();
     const width = Math.ceil(renderer.domElement.clientWidth * pixelRatio);
     const height = Math.ceil(renderer.domElement.clientHeight * pixelRatio);
-    PointCloudOctreePicker.updatePickRenderTarget(this.pickState, width, height);
+    PointCloudOctreePickerHelper.updatePickRenderTarget(this.pickState, width, height);
 
     const pixelPosition = PointCloudOctreePicker.helperVec3; // Use helper vector to prevent extra allocations.
 
@@ -100,9 +58,9 @@ export class PointCloudOctreePicker {
     const x = Math.floor(clamp(pixelPosition.x - halfPickWndSize, 0, width));
     const y = Math.floor(clamp(pixelPosition.y - halfPickWndSize, 0, height));
 
-    PointCloudOctreePicker.prepareRender(renderer, x, y, pickWndSize, pickMaterial, pickState);
+    PointCloudOctreePickerHelper.prepareRender(renderer, x, y, pickWndSize, pickMaterial, pickState);
 
-    const renderedNodes = PointCloudOctreePicker.render(
+    const renderedNodes = PointCloudOctreePickerHelper.render(
       renderer,
       camera,
       pickMaterial,
@@ -116,287 +74,8 @@ export class PointCloudOctreePicker {
     pickMaterial.clearVisibleNodeTextureOffsets();
 
     // Read back image and decode hit point
-    const pixels = PointCloudOctreePicker.readPixels(renderer, x, y, pickWndSize);
-    const hit = PointCloudOctreePicker.findHit(pixels, pickWndSize, renderedNodes, camera);
-    return PointCloudOctreePicker.getPickPoint(hit, renderedNodes);
-  }
-
-  private static prepareRender(
-    renderer: WebGLRenderer,
-    x: number,
-    y: number,
-    pickWndSize: number,
-    pickMaterial: PointCloudMaterial,
-    pickState: IPickState
-  ) {
-    // Render the intersected nodes onto the pick render target, clipping to a small pick window.
-    renderer.setScissor(x, y, pickWndSize, pickWndSize);
-    renderer.setScissorTest(true);
-    renderer.state.buffers.depth.setTest(pickMaterial.depthTest);
-    renderer.state.buffers.depth.setMask(pickMaterial.depthWrite);
-    renderer.state.setBlending(NoBlending);
-
-    renderer.setRenderTarget(pickState.renderTarget);
-
-    // Save the current clear color and clear the renderer with black color and alpha 0.
-    renderer.getClearColor(this.clearColor);
-    const oldClearAlpha = renderer.getClearAlpha();
-    renderer.setClearColor(COLOR_BLACK, 0);
-    renderer.clear(true, true, true);
-    renderer.setClearColor(this.clearColor, oldClearAlpha);
-  }
-
-  private static render(
-    renderer: WebGLRenderer,
-    camera: Camera,
-    pickMaterial: PointCloudMaterial,
-    octrees: PointCloudOctree[],
-    ray: Ray,
-    pickState: IPickState,
-    params: Partial<PickParams>
-  ): RenderedNode[] {
-    const renderedNodes: RenderedNode[] = [];
-    for (const octree of octrees) {
-      // Get all the octree nodes which intersect the picking ray. We only need to render those.
-      const nodes = PointCloudOctreePicker.nodesOnRay(octree, ray);
-      if (!nodes.length) {
-        continue;
-      }
-
-      PointCloudOctreePicker.updatePickMaterial(pickMaterial, octree.material, params);
-      pickMaterial.updateMaterial(octree, nodes, camera, renderer);
-
-      if (params.onBeforePickRender) {
-        params.onBeforePickRender(pickMaterial, pickState.renderTarget);
-      }
-
-      // Create copies of the nodes so we can render them differently than in the normal point cloud.
-      pickState.scene.children = PointCloudOctreePicker.createTempNodes(
-        octree,
-        nodes,
-        pickMaterial,
-        renderedNodes.length
-      );
-
-      renderer.render(pickState.scene, camera);
-
-      nodes.forEach(node => renderedNodes.push({ node, octree }));
-    }
-    return renderedNodes;
-  }
-
-  private static nodesOnRay(octree: PointCloudOctree, ray: Ray): IPointCloudTreeNode[] {
-    const nodesOnRay: IPointCloudTreeNode[] = [];
-
-    const rayClone = ray.clone();
-    for (const node of octree.visibleNodes) {
-      const sphere = PointCloudOctreePicker.helperSphere.copy(node.boundingSphere).applyMatrix4(octree.matrixWorld);
-
-      if (rayClone.intersectsSphere(sphere)) {
-        nodesOnRay.push(node);
-      }
-    }
-
-    return nodesOnRay;
-  }
-
-  private static readPixels(renderer: WebGLRenderer, x: number, y: number, pickWndSize: number): Uint8Array {
-    // Read the pixel from the pick render target.
-    const pixels = new Uint8Array(4 * pickWndSize * pickWndSize);
-    renderer.readRenderTargetPixels(renderer.getRenderTarget()!, x, y, pickWndSize, pickWndSize, pixels);
-    renderer.setScissorTest(false);
-    renderer.setRenderTarget(null!);
-    return pixels;
-  }
-
-  private static createTempNodes(
-    octree: PointCloudOctree,
-    nodes: IPointCloudTreeNode[],
-    pickMaterial: PointCloudMaterial,
-    nodeIndexOffset: number
-  ): Points[] {
-    const tempNodes: Points[] = [];
-    for (let i = 0; i < nodes.length; i++) {
-      const node = nodes[i];
-      const sceneNode = node.sceneNode;
-      const tempNode = new Points(sceneNode.geometry, pickMaterial);
-      tempNode.matrix = sceneNode.matrix;
-      tempNode.matrixWorld = sceneNode.matrixWorld;
-      tempNode.matrixAutoUpdate = false;
-      tempNode.frustumCulled = false;
-      const nodeIndex = nodeIndexOffset + i + 1;
-      if (nodeIndex > 255) {
-        console.error('More than 255 nodes for pick are not supported.');
-      }
-      tempNode.onBeforeRender = PointCloudMaterial.makeOnBeforeRender(octree, node, nodeIndex);
-
-      tempNodes.push(tempNode);
-    }
-    return tempNodes;
-  }
-
-  private static updatePickMaterial(
-    pickMaterial: PointCloudMaterial,
-    nodeMaterial: PointCloudMaterial,
-    params: Partial<PickParams>
-  ): void {
-    pickMaterial.pointSizeType = nodeMaterial.pointSizeType;
-    pickMaterial.shape = nodeMaterial.shape;
-    pickMaterial.size = nodeMaterial.size;
-    pickMaterial.minSize = nodeMaterial.minSize;
-    pickMaterial.maxSize = nodeMaterial.maxSize;
-    pickMaterial.classification = nodeMaterial.classification;
-    pickMaterial.useFilterByNormal = nodeMaterial.useFilterByNormal;
-    pickMaterial.filterByNormalThreshold = nodeMaterial.filterByNormalThreshold;
-
-    if (params.pickOutsideClipRegion) {
-      pickMaterial.clipMode = ClipMode.DISABLED;
-    } else {
-      pickMaterial.clipMode = nodeMaterial.clipMode;
-      pickMaterial.setClipBoxes(nodeMaterial.clipMode === ClipMode.CLIP_OUTSIDE ? nodeMaterial.clipBoxes : []);
-    }
-  }
-
-  private static updatePickRenderTarget(pickState: IPickState, width: number, height: number): void {
-    if (pickState.renderTarget.width === width && pickState.renderTarget.height === height) {
-      return;
-    }
-
-    pickState.renderTarget.dispose();
-    pickState.renderTarget = PointCloudOctreePicker.makePickRenderTarget();
-    pickState.renderTarget.setSize(width, height);
-  }
-
-  private static makePickRenderTarget() {
-    return new WebGLRenderTarget(1, 1, {
-      minFilter: LinearFilter,
-      magFilter: NearestFilter,
-      format: RGBAFormat
-    });
-  }
-
-  private static findHit(
-    pixels: Uint8Array,
-    pickWndSize: number,
-    nodes: RenderedNode[],
-    camera: THREE.Camera
-  ): PointCloudHit | null {
-    const ibuffer = new Uint32Array(pixels.buffer);
-
-    // Find closest hit inside pixelWindow boundaries and closest to the camera.
-    let minScreen = Number.MAX_VALUE;
-    let minCameraDistance = Number.MAX_VALUE;
-    let hit: PointCloudHit | null = null;
-    for (let u = 0; u < pickWndSize; u++) {
-      for (let v = 0; v < pickWndSize; v++) {
-        const offset = u + v * pickWndSize;
-        const screenDistance = Math.pow(u - (pickWndSize - 1) / 2, 2) + Math.pow(v - (pickWndSize - 1) / 2, 2);
-
-        const pcIndex = pixels[4 * offset + 3];
-        pixels[4 * offset + 3] = 0;
-        const pIndex = ibuffer[offset];
-
-        if (pcIndex > 0 && pcIndex !== 255 && screenDistance < minScreen) {
-          const points = nodes[pcIndex - 1]?.node.sceneNode;
-          this.helperVec3
-            .fromBufferAttribute(points.geometry.attributes['position'], pIndex)
-            .applyMatrix4(points.matrixWorld);
-          const distanceToCamera = this.helperVec3.distanceToSquared(camera.position);
-
-          if (distanceToCamera < minCameraDistance) {
-            hit = {
-              pIndex: pIndex,
-              pcIndex: pcIndex - 1
-            };
-
-            minScreen = screenDistance;
-            minCameraDistance = distanceToCamera;
-          }
-        }
-      }
-    }
-    return hit;
-  }
-
-  private static getPickPoint(hit: PointCloudHit | null, nodes: RenderedNode[]): PickPoint | null {
-    if (!hit) {
-      return null;
-    }
-
-    const points = nodes[hit.pcIndex]?.node.sceneNode;
-    if (!points) {
-      return null;
-    }
-
-    const point: PickPoint = { pointIndex: hit.pIndex, object: points, position: new Vector3() };
-
-    point.pointCloud = nodes[hit.pcIndex].octree;
-
-    const attributes = points.geometry.attributes;
-
-    for (const property in attributes) {
-      if (!attributes.hasOwnProperty(property)) {
-        continue;
-      }
-
-      const values = attributes[property] as BufferAttribute;
-
-      // tslint:disable-next-line:prefer-switch
-      if (property === 'position') {
-        PointCloudOctreePicker.addPositionToPickPoint(point, hit, values, points);
-      } else if (property === 'normal') {
-        PointCloudOctreePicker.addNormalToPickPoint(point, hit, values, points);
-      } else if (property === 'indices') {
-        // TODO
-      } else {
-        if (values.itemSize === 1) {
-          point[property] = values.array[hit.pIndex];
-        } else {
-          const value: number[] = [];
-          for (let j = 0; j < values.itemSize; j++) {
-            value.push(values.array[values.itemSize * hit.pIndex + j]);
-          }
-          point[property] = value;
-        }
-      }
-    }
-
-    return point;
-  }
-
-  private static addPositionToPickPoint(
-    point: PickPoint,
-    hit: PointCloudHit,
-    values: BufferAttribute,
-    points: Points
-  ): void {
-    point.position = new Vector3().fromBufferAttribute(values, hit.pIndex).applyMatrix4(points.matrixWorld);
-  }
-
-  private static addNormalToPickPoint(
-    point: PickPoint,
-    hit: PointCloudHit,
-    values: BufferAttribute,
-    points: Points
-  ): void {
-    const normal = new Vector3().fromBufferAttribute(values, hit.pIndex);
-    const normal4 = new Vector4(normal.x, normal.y, normal.z, 0).applyMatrix4(points.matrixWorld);
-    normal.set(normal4.x, normal4.y, normal4.z);
-
-    point.normal = normal;
-  }
-
-  private static getPickState() {
-    const scene = new Scene();
-    scene.autoUpdate = false;
-
-    const material = new PointCloudMaterial();
-    material.pointColorType = PotreePointColorType.PointIndex;
-
-    return {
-      renderTarget: PointCloudOctreePicker.makePickRenderTarget(),
-      material: material,
-      scene: scene
-    };
+    const pixels = PointCloudOctreePickerHelper.readPixels(renderer, x, y, pickWndSize);
+    const hit = PointCloudOctreePickerHelper.findHit(pixels, pickWndSize, renderedNodes, camera);
+    return PointCloudOctreePickerHelper.getPickPoint(hit, renderedNodes);
   }
 }
